move poses to camera panel
