+++ src/ui/editor.ts
@@ -7,6 +7,7 @@ import { ViewCube } from './view-cube';
 import { Menu } from './menu';
 import { ScenePanel } from './scene-panel';
 import { ViewPanel } from './view-panel';
+import { CameraPanel } from './camera-panel';
 import { BottomToolbar } from './bottom-toolbar';
 import { RightToolbar } from './right-toolbar';
 import { ModeToggle } from './mode-toggle';
@@ -82,6 +83,7 @@ class EditorUI {
         // bottom toolbar
         const scenePanel = new ScenePanel(events, tooltips);
         const viewPanel = new ViewPanel(events, tooltips);
+        const cameraPanel = new CameraPanel(events, tooltips);
         const bottomToolbar = new BottomToolbar(events, tooltips);
         const rightToolbar = new RightToolbar(events, tooltips);
         const modeToggle = new ModeToggle(events, tooltips);
@@ -92,6 +94,7 @@ class EditorUI {
         canvasContainer.append(toolsContainer);
         canvasContainer.append(scenePanel);
         canvasContainer.append(viewPanel);
+        canvasContainer.append(cameraPanel);
         canvasContainer.append(bottomToolbar);
         canvasContainer.append(rightToolbar);
         canvasContainer.append(modeToggle);
