+++ src/ui/scss/colors.scss
@@ -14,5 +14,6 @@ $clr-disabled: #7c7678;
 $clr-active: white;
 $clr-hilight: #f60;
 
+$bcg-lighter: #444;
 $bcg-light: #555;
 $bcg-darkest: #181818;
