+++ src/ui/scss/panel.scss
@@ -51,7 +51,7 @@
             height: 28px;
             padding: 0px;
             margin: 0px;
-            background-color: $bcg-primary;
+            background-color: $bcg-lighter;
         }
     }    
 }
