+++ src/ui/scss/style.scss
@@ -6,6 +6,7 @@
 @use 'menu.scss';
 @use 'scene-panel.scss';
 @use 'view-panel.scss';
+@use 'camera-panel.scss';
 @use 'splat-list.scss';
 @use 'transform.scss';
 @use 'bottom-toolbar.scss';
