+++ src/ui/scss/view-panel.scss
@@ -64,37 +64,4 @@
             }
         }
     }
-
-    > .view-panel-list-container {
-        width: 100%;
-        min-height: 80px;
-        max-height: 220px;
-
-        padding: 4px 0px;
-
-        overflow-y: auto;
-
-        > .view-panel-list {
-            width: 100%;
-            display: flex;
-            flex-direction: column;
-
-            > .view-panel-list-row {
-                display: flex;
-                flex-direction: row;
-                flex-shrink: 0;
-                flex-grow: 0;
-
-                cursor: pointer;
-
-                > .view-panel-list-row-text {
-                    flex-grow: 1;
-                }
-            }
-        }
-
-        .selected {
-            background-color: $bcg-darker;
-        }
-    }
 }
